// pruebas/01-reading-list/KrlosPK/src/app/library/interfaces/types.d.ts
// Generated by https://quicktype.io

export interface Books {
  book: Book;
}

export interface Book {
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: Author;
}

export interface Author {
  name: string;
  otherBooks: string[];
}